+++ src/KalitkaAgent.Windows/Program.cs
@@ -26,6 +26,7 @@ builder.Services.AddHttpClient<CoreClient>((sp, http) =>
 // enforcer that adds on redeem / removes on expiry.
 builder.Services.AddSingleton(TimeProvider.System);
 builder.Services.AddSingleton<ILocalGroup, WindowsLocalGroup>();
+builder.Services.AddSingleton<ISessionKiller, WtsSessionKiller>();
 builder.Services.AddSingleton(sp =>
     new RdpJournal(sp.GetRequiredService<Microsoft.Extensions.Options.IOptions<AgentConfig>>().Value.RdpJournalPath));
 builder.Services.AddSingleton<RdpEnforcer>();